fix:tests for argumentsCheckFn
